H3DAPI warning fixes (vs2010 64-bit)
Summary:
Additional fixes for warnings that only show up when doing 64-bit builds (primarily because size_t and int are no longer the same size in 64-bit visual studio)

Related task: T712

Test Plan: Will run unit tests which should open not only all the directly edited nodes but also many of the nodes that might be indirectly affected by any of the changes.

Reviewers: markus

Reviewed By: markus

Subscribers: michaels, markus

Tags: #h3dapi_2.x

Differential Revision: https://www.h3dapi.org:4000/D33

+++ trunk/H3DAPI/src/Appearance.cpp
@@ -359,7 +359,7 @@ void Appearance::traverseSG( TraverseInfo &ti ) {
     glGetIntegerv( GL_MAX_LIGHTS, &max_lights );
   }
   NavigationInfo *ni = NavigationInfo::getActive();
-  unsigned int nr_lights = ti.getActiveLightNodes().size();
+  size_t nr_lights = ti.getActiveLightNodes().size();
   if( !ni || ni->headlight->getValue() ) {
     ++nr_lights;
   }
